feat(controller): handle stok and jumlah logic in obat and periksa pasien controllers

+++ app/Http/Controllers/ObatController.php
@@ -21,12 +21,14 @@ class ObatController extends Controller
             'nama_obat' => 'required|string',
             'kemasan' => 'required|string',
             'harga' => 'required|integer',
+            'stok' => 'required|integer|min:0',
         ]);
 
         Obat::create([
             'nama_obat' => $request->nama_obat,
             'kemasan' => $request->kemasan,
-            'harga' => $request->harga
+            'harga' => $request->harga,
+            'stok' => $request->stok
         ]);
 
         return redirect()->route('obat.index')
@@ -46,13 +48,15 @@ class ObatController extends Controller
             'nama_obat' => 'required|string',
             'kemasan' => 'nullable|string',
             'harga' => 'required|integer',
+            'stok' => 'required|integer|min:0',
         ]);
 
         $obat = Obat::findOrFail($id);
         $obat->update([
             'nama_obat' => $request->nama_obat,
             'kemasan' => $request->kemasan,
-            'harga' => $request->harga
+            'harga' => $request->harga,
+            'stok' => $request->stok
         ]);
 
         return redirect()->route('obat.index')
@@ -62,8 +66,18 @@ class ObatController extends Controller
 
     public function destroy(string $id){
         $obat = Obat::findOrFail($id);
+    
+        // VALIDASI TAMBAHAN: CEK KETERGANTUNGAN DATA
+        // Cek apakah obat ini ada di tabel detail_periksa?
+        
+        if ($obat->detailPeriksas()->count() > 0) {
+            return redirect()->route('obat.index')
+                ->with('message', 'GAGAL! Obat tidak bisa dihapus karena sudah pernah digunakan dalam riwayat pemeriksaan pasien. Silahkan edit stok menjadi 0 saja.')
+                ->with('type', 'danger'); 
+        }
+    
         $obat->delete();
-
+    
         return redirect()->route('obat.index')
             ->with('message','Data Obat berhasil di Hapus')
             ->with('type','success');
